+++ src/components/style/Style.scss
@@ -1,4 +1,5 @@
 @import 'SingUp-LogIn';
+@import 'Menu';
 
 body{
     background-color: rgba(61, 68, 244, 0.65);
