+++ src/components/style/_Menu.scss
@@ -1,0 +1,14 @@
+.product{
+    background-color: rgba(217, 162, 150, 0.63);
+    display: flex;
+    
+    .image{
+        background-color: rgba(247, 231, 208, 0.4);
+        width: 5vmin;
+        height: 5vmin;
+        img{
+           width: 80%;
+        }
+    }
+
+}
